requirements.txt
asgiref==3.6.0
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
idna==3.4
Markdown==3.3.4
martor==1.6.26
Pillow==9.5.0
PyJWT==2.7.0
pymongo==3.12.3
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==2.0.2
webencodings==0.5.1
